Use $CATKIN_WS instead of ~/catkin_ws
On the one hand `$CATKIN_WS` is defined to `/root/catkin_ws`, on the other hand at some places `~/catkin_ws` is used.
That works fine as long as it is run by root, but doesn't work if you run this script with sudo (e.g. to run the CI in a local docker environment in order to debug the build job).